// src/lib/openapi/squad/models/UpdateWorkspacePayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Squad API
 * API for managing Squad resources
 *
 * The version of the OpenAPI document: 4.0.0
 * Contact: help@meetsquad.ai
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime.js';
/**
 * Request schema for updating a workspace
 * @export
 * @interface UpdateWorkspacePayload
 */
export interface UpdateWorkspacePayload {
    /**
     * Name of the workspace
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    homepageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    logoUrl?: string | null;
    /**
     * Mission statement for the workspace
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    missionStatement?: string;
    /**
     * Detailed description of the workspace
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    description?: string;
    /**
     * Status of the workspace
     * @type {string}
     * @memberof UpdateWorkspacePayload
     */
    status?: UpdateWorkspacePayloadStatusEnum;
}


/**
 * @export
 */
export const UpdateWorkspacePayloadStatusEnum = {
    Initializing: 'Initializing',
    OnboardingDeepResearch: 'OnboardingDeepResearch',
    OnboardingConsiderWorkspace: 'OnboardingConsiderWorkspace',
    OnboardingConsiderOpportunities: 'OnboardingConsiderOpportunities',
    OnboardingConsiderOpportunitiesToOutcomes: 'OnboardingConsiderOpportunitiesToOutcomes',
    OnboardingConsiderSolutionGeneration: 'OnboardingConsiderSolutionGeneration',
    OnboardingConsiderSolutionsToBuild: 'OnboardingConsiderSolutionsToBuild',
    OnboardingSelectSolutionsToBuild: 'OnboardingSelectSolutionsToBuild',
    OnboardingMapSolutionsToRoadmap: 'OnboardingMapSolutionsToRoadmap',
    Active: 'Active',
    Inactive: 'Inactive'
} as const;
export type UpdateWorkspacePayloadStatusEnum = typeof UpdateWorkspacePayloadStatusEnum[keyof typeof UpdateWorkspacePayloadStatusEnum];


/**
 * Check if a given object implements the UpdateWorkspacePayload interface.
 */
export function instanceOfUpdateWorkspacePayload(value: object): value is UpdateWorkspacePayload {
    return true;
}

export function UpdateWorkspacePayloadFromJSON(json: any): UpdateWorkspacePayload {
    return UpdateWorkspacePayloadFromJSONTyped(json, false);
}

export function UpdateWorkspacePayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateWorkspacePayload {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'homepageUrl': json['homepageUrl'] == null ? undefined : json['homepageUrl'],
        'logoUrl': json['logoUrl'] == null ? undefined : json['logoUrl'],
        'missionStatement': json['missionStatement'] == null ? undefined : json['missionStatement'],
        'description': json['description'] == null ? undefined : json['description'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function UpdateWorkspacePayloadToJSON(json: any): UpdateWorkspacePayload {
    return UpdateWorkspacePayloadToJSONTyped(json, false);
}

export function UpdateWorkspacePayloadToJSONTyped(value?: UpdateWorkspacePayload | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'homepageUrl': value['homepageUrl'],
        'logoUrl': value['logoUrl'],
        'missionStatement': value['missionStatement'],
        'description': value['description'],
        'status': value['status'],
    };
}